+++ lib/src/platform_specific/file_manager/file_manager_stub.dart
@@ -1,0 +1,4 @@
+import 'file_manager.dart';
+
+PlatformFileManager getFileManager() =>
+    throw UnimplementedError("File Picker is Not Implementd in current platform");